+++ TOMB5/specific/winmain.cpp
@@ -102,7 +102,7 @@ void WinDisplayString(long x, long y, char* string, ...)
 	va_start(list, string);
 	vsprintf(buf, string, list);
 	va_end(list);
-	PrintString((ushort)x, (ushort)y, 6, buf, 0);
+	PrintString(x, y, 6, buf, 0);
 }
 
 void WinProcMsg()
